# perf: Amortize clearing unsorted cache entries (Juno genesis fix)