--- a/store/cachekv/search_benchmark_test.go
+++ b/store/cachekv/search_benchmark_test.go
@@ -8,29 +8,36 @@ import (
 
 func BenchmarkLargeUnsortedMisses(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		cache := map[string]*cValue{}
-		unsorted := map[string]struct{}{}
-		for i := 0; i < 100_000; i++ {
-			key := "A" + strconv.Itoa(i)
-			unsorted[key] = struct{}{}
-			cache[key] = &cValue{}
-		}
+		b.StopTimer()
+		store := generateStore()
+		b.StartTimer()
 
-		for i := 0; i < 100_000; i++ {
-			key := "Z" + strconv.Itoa(i)
-			unsorted[key] = struct{}{}
-			cache[key] = &cValue{}
-		}
-
-		store := Store{
-			cache:         cache,
-			unsortedCache: unsorted,
-			sortedCache:   db.NewMemDB(),
-		}
 		for k := 0; k < 10000; k++ {
 			// cache has A + Z values
 			// these are within range, but match nothing
 			store.dirtyItems([]byte("B1"), []byte("B2"))
 		}
+	}
+}
+
+func generateStore() *Store {
+	cache := map[string]*cValue{}
+	unsorted := map[string]struct{}{}
+	for i := 0; i < 5000; i++ {
+		key := "A" + strconv.Itoa(i)
+		unsorted[key] = struct{}{}
+		cache[key] = &cValue{}
+	}
+
+	for i := 0; i < 5000; i++ {
+		key := "Z" + strconv.Itoa(i)
+		unsorted[key] = struct{}{}
+		cache[key] = &cValue{}
+	}
+
+	return &Store{
+		cache:         cache,
+		unsortedCache: unsorted,
+		sortedCache:   db.NewMemDB(),
 	}
 }