@blazeroni 2022-08-10
I've updated the PR to a realistic implementation.  It adds a module-internal interface which allows the values to be cached.