@blazeroni 2022-08-11
I dug deeper and found the root problem.  Replaced most of the code for a real fix.