@blazeroni 2022-08-12
Added a benchmark which illustrates the problem.  Finishes in a few seconds with the changes in the PR.  Takes minutes without it.