@blazeroni 2022-08-16
@odeke-em thanks for the feedback!  I've updated the benchmark and added it as a comment.  I had to reduce the size of the original benchmark (from 200k to 10k unsorted values) since the before benchmark was crashing with values much higher than what I used.  Regardless, the difference between before/after is still extreme.