@joeabbey 2022-08-16
Impressive results.  I've measured similar duration for crisis invariants on similar hardware.  Moreover since this PR changes CacheKV performance, the whole startup from a large `genesis.json` improves.

Without this change the invariants took roughly 14 hours to complete, are there was a stage of replaying blocks and other initialization that took roughly two hours.  Validators were reporting 16+ hours to begin dialing peers.

With this change the entire initialization process completes in around 5 minutes.