@blazeroni 2022-08-16
@kocubinski yes, your analysis matches my own.

To test/verify this, I used a modified version of the latest Juno production release, available here: https://github.com/blazeroni/juno-bounty/commits/v9.0.0-invariants.  That version depends on a modified v0.45.6 Cosmos SDK, with the only change being the changes in this PR.  Instructions to obtain the Juno genesis file are here: https://github.com/CosmosContracts/incident-response/blob/main/28-July-22/genesis.md.  You'll just need to install, init, replace genesis, and start.

Cheers!